backend/requirements.txt
dj-database-url==0.4.2
Django==1.11.6
django-cors-headers==2.1.0
django-environ==0.4.4
django-rest-multiple-models==1.8.2
djangorestframework==3.7.1
drf-tracking==1.3.1
psycopg2==2.7.4
pytz==2017.2
six==1.11.0
uptime==3.0.1
